Refactor ListWithProperty arguments
(breaking change)

In order for delete snapshots to work when we set the delete-at property on its parent filesystem, we need to be able to set the property scope on the function.
All of the arguments except property can also be optional now. Use a struct for better forwards compatibility in the future.